MLX: reject non-uint32 width; move sample to custom_ops; docs
